SearchBot/requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
textwrap
google.generativeai
IPython